test-requirements.txt
asynctest
codecov>=1.4.0
coverage>=4.0.3
flake8
isort
mock>=2.0.0
nose>=1.3.7
pluggy>=0.3.1
py>=1.4.31
pytest
pytest-aiohttp
pytest-xdist
randomize>=0.13
recommonmark
sphinx>=1.2.1,!=1.3b1,<1.4 # BSD
